Renamed `PresenterFactory` namespace to `Instancing`
